Create quote_to_image.php
turns a specified csv file in the same folder as this script (litclock_annotated_expanded.csv) into a series of images with and without metadata. Needs the image and metadata folder created for it.

You need some dependencies, need to enable gd extension and download/enable Image magic (Imagick) extension also. 

This code runs in windows and the fonts listed here need to be in the same folder as the script itself - you can get it to run to reference the default fonts folder but I never got this working.